requirements.txt
fastapi
uvicorn
motor
pymongo
pytest
httpx
python-dotenv
pydantic